requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-celery-beat==2.5.0
networkx==3.2.1
pandas==2.1.3
numpy==1.25.2
pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0